requirements.txt
boto==2.49.0
boto3==1.9.61
botocore==1.12.61
Django==2.1.4
django-storages==1.7.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docutils==0.14
jmespath==0.9.3
Pillow==5.3.0
PyJWT==1.7.0
python-dateutil==2.7.5
pytz==2018.7
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
